requirements.txt
# sqladmin 核心依赖
sqladmin[full]>=0.16.0

# SQLAlchemy 数据库依赖
sqlalchemy>=2.0.0
alembic>=1.12.0  # 数据库迁移工具

# 数据库驱动
asyncpg>=0.29.0  # PostgreSQL 异步驱动
psycopg2-binary>=2.9.0  # PostgreSQL 同步驱动
aiomysql>=0.2.0  # MySQL
aiosqlite>=0.19.0  # SQLite

# 密码哈希
bcrypt>=4.0.0

# FastAPI 依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 表单处理
WTForms>=3.1.0

# 配置管理
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 环境变量管理
python-dotenv>=1.0.0

# JSON处理增强
python-json-logger>=2.0.0

# 时间处理
python-dateutil>=2.8.0

# 开发依赖（可选）
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# isort>=5.12.0

# 工具库
python-dotenv
pydantic
pydantic-settings
